Created a rough version of the UI
I created a rough functional version of the UI that we planned out. The new starting menu and choice menu have all their functionalities and I also implemented the ability to choose different puzzles to solve. The looks and visuals are still in the rough draft phase but the general layout is pretty much there.