More visuals along with other code
I took Yanitsa's visuals and continued to work with it, mainly making it so that the images on the buttons covered the entire button and also tweaked the random game button to make it so that it also randomly chose the number of pieces in the game. There is also other code that is included from my attempts to figure out how to make puzzle pieces that clip together like in most jigsaw games. That code is still very much a work in progress.