Added debug toggle to menu, added load key, added fix for alignment
I added a debug toggle to the first menu in the new_menu scene that is activated by pressing the d key. I also added an l key press in jigsaw_puzzle_1 in order to make it easier for people to load from a database, though they will need to implement the actual functionality. I also made some changes to help with alignment with regards to using timers, but there are still weird glitches where there are phantom snaps, where the sound plays or the pieces jitter and then settle. Some sounds were added but they are still very much a work in progress as some work while others do not.